@@ -1,3 +1,3 @@
 module vftpd
 
-go 1.21
+go 1.22
